+++ src/util.cpp
@@ -193,7 +193,7 @@ void set_process_scheduler(const int& pid, const int& policy_index, const int& p
   sched_setscheduler(pid, policy_index, &policy_params);
 }
 
-void set_sched_runtime(const int& pid, const uint64_t& value, const int& policy_index, const uint& flags) {
+void set_sched_runtime(const int& pid, const double& value, const int& policy_index, const uint& flags) {
   sched_attr attr = {};
 
   attr.size = sizeof(sched_attr);
