+++ src/util.hpp
@@ -39,7 +39,7 @@ void clear_cpu_affinity(const int& pid);
 
 void set_process_scheduler(const int& pid, const int& policy_index, const int& priority);
 
-void set_sched_runtime(const int& pid, const uint64_t& value, const int& policy_index, const uint& flags);
+void set_sched_runtime(const int& pid, const double& value, const int& policy_index, const uint& flags);
 
 auto get_sched_runtime(const int& pid, const uint& flags) -> uint64_t;
